#3031: Rename src into shotgunEvents
