+++ shotgunEvents/examplePlugins/__init__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/python2.7
+"""
+    Set environment and run software using Rez
+    Need REZ_PACKAGES_PATH and REZ_CONFIG _PATH to be set
+"""
+
+# BASE IMPORTS
+# ====================
+
+
+
+
+if __name__ == '__main__':
+    pass
