+++ shotgunEvents/examplePlugins/calculateCutDuration.py
