+++ shotgunEvents/examplePlugins/logArgs.py
