+++ shotgunEvents/examplePlugins/sharedStateA.py
