+++ shotgunEvents/examplePlugins/sharedStateB.py
