+++ shotgunEvents/examplePlugins/sharedStateC.py
