+++ shotgunEvents/scripts/createPluginManagerSGEntries.py
