+++ shotgunEvents/shotgunEventDaemon.conf.example
